Update README.md
SolarSynapse is a modular device with dual-OS support for Windows/Linux and Android on a mini ITX motherboard. It features solar power and power bank integration for sustainable use. Designed for upgradability, it adapts to your needs, minimizing the need for multiple gadgets. A future-proof solution for eco-conscious tech enthusiasts.

ومتنقل